Changed competitions view
