+++ mastersofcode-adminclient/src/main/webapp/WEB-INF/views/competitions/competitions.twig
@@ -1,0 +1,34 @@
+{% extends "dashboard/index.twig" %}
+
+{% block title %}
+    {{ labels.producer_dashboard_index_page_title }}
+{% endblock %}
+
+{% block heading %}
+    <div class="row">
+        <div class="col-lg-12">
+            <h1 class="page-header">
+                Wedstrijden
+            </h1>
+        </div>
+    </div>
+{% endblock %}
+
+{% block content %}
+    <div class="row">
+        <div class="col-lg-12">
+            <p class="error">            
+            </p>
+            {% if competitions is empty%}
+                <p> Er zijn nog geen competities </p>
+            {% else %}
+                
+                {% for competition in competitions %}
+
+
+
+                {% endfor %}
+            {% endif %}
+        </div>
+    </div>
+{% endblock %}
