--- mastersofcode-adminclient/src/main/webapp/WEB-INF/views/dashboard/competitions.twig
@@ -1,43 +1,0 @@
-{% extends "dashboard/index.twig" %}
-
-{% block title %}
-    {{ labels.producer_dashboard_index_page_title }}
-{% endblock %}
-
-{% block heading %}
-<div class="row">
-    <div class="col-lg-12">
-        <h1 class="page-header">
-            Wedstrijden
-        </h1>
-    </div>
-</div>
-{% endblock %}
-
-{% block content %}
-<div class="row">
-    <div class="col-lg-12">
-        <p class="error">            
-        </p>
-        {% if compitions is not empty %}            
-            <table class="table">
-                {% for competition in competitions %}
-                    <tr>
-                        <td>
-                            <a href="/mok/competitions/{{ competition.id }}">
-                                {{ competition.title }}
-                            </a>
-                        </td>
-                    </tr>
-
-                {% endfor %}              
-            </table>
-        {% else %}
-            Nog geen wedstrijden gepubliceerd.
-            <a href="/mok/competitions/add">
-                Voeg toe...
-            </a>
-        {% endif %}
-    </div>
-</div>
-{% endblock %}
