+++ mastersofcode-adminclient/src/main/webapp/WEB-INF/views/dashboard/index.twig
@@ -30,19 +30,19 @@
             <div class="collapse navbar-collapse navbar-ex1-collapse">
                 <ul class="nav navbar-nav side-nav">
                     <li>    
-                        <a href="wedstrijden"><i class="fa fa-gamepad"></i> Wedstrijden </a>
+                        <a href="/mok/competitions"><i class="fa fa-gamepad"></i> Wedstrijden </a>
                     </li>
 
                     <li>    
-                        <a href="teams"><i class="fa fa-users"></i> Teams </a>
+                        <a href="/mok/teams"><i class="fa fa-users"></i> Teams </a>
                     </li>
    
                     <li>    
-                        <a href="deelnemers"><i class="fa fa-user"></i> Deelnemers </a>
+                        <a href="/mok/members"><i class="fa fa-user"></i> Deelnemers </a>
                     </li>
 
                     <li>    
-                        <a href="opdrachten"><i class="fa fa-file-text-o"></i> Opdrachten </a>
+                        <a href="/mok/assignments"><i class="fa fa-file-text-o"></i> Opdrachten </a>
                     </li>
 
                     <li>
